# src/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Discount Service API",
    "version": "1.0.0",
    "description": "API for managing vouchers, promotions, and applying discounts to orders",
    "contact": {
      "name": "API Support"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    }
  ],
...
  "components": {
    "schemas": {
      "Voucher": {
        "type": "object",
        "required": [
          "discountType",
          "discountValue",
          "expirationDate",
          "usageLimit"
        ],
        "properties": {
          "code": {
            "type": "string",
            "description": "Voucher code (auto-generated if not provided)",
            "example": "SAVE20"
          },
          "discountType": {
            "type": "string",
            "enum": [
              "percentage",
              "fixed"
            ],
            "description": "Type of discount",
            "example": "percentage"
          },
          "discountValue": {
            "type": "number",
            "description": "Discount value (percentage or fixed amount)",
            "example": 20
          },
          "expirationDate": {
            "type": "string",
            "format": "date-time",
            "description": "Expiration date",
            "example": "2024-12-31T23:59:59Z"
          },
          "usageLimit": {
            "type": "integer",
            "description": "Maximum number of times the voucher can be used",
            "example": 100
          },
          "minimumOrderValue": {
            "type": "number",
            "description": "Minimum order value required to use this voucher",
            "example": 50
          },
          "isActive": {
            "type": "boolean",
            "description": "Whether the voucher is active",
            "example": true
          }
        }
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "Error message"
          },
          "details": {
            "type": "array",
            "items": {
              "type": "object"
            },
            "description": "Validation error details"
          }
        }
      }
    }
  },
  "paths": {
...
    "/api/vouchers": {
      "get": {
        "summary": "Get all vouchers",
        "tags": [
          "Vouchers"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "isActive",
            "schema": {
              "type": "boolean"
            },
            "description": "Filter by active status"
          }
        ],
        "responses": {
          "200": {
            "description": "List of vouchers",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Voucher"
                      }
                    },
                    "count": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new voucher",
        "tags": [
          "Vouchers"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Voucher"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Voucher created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Voucher"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Voucher code already exists"
          },
          "429": {
            "description": "Too many requests",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RateLimitError"
                }
              }
            }
          }
        }
      }
    },
    "/api/vouchers/{id}": {
      "get": {
        "summary": "Get voucher by ID",
        "tags": [
          "Vouchers"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Voucher ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Voucher details",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Voucher"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Voucher not found"
          }
        }
      },
      "put": {
        "summary": "Update a voucher",
        "tags": [
          "Vouchers"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Voucher ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "code": {
                    "type": "string"
                  },
                  "discountType": {
                    "type": "string",
                    "enum": [
                      "percentage",
                      "fixed"
                    ]
                  },
                  "discountValue": {
                    "type": "number"
                  },
                  "expirationDate": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "usageLimit": {
                    "type": "integer"
                  },
                  "minimumOrderValue": {
                    "type": "number"
                  },
                  "isActive": {
                    "type": "boolean"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Voucher updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Voucher"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Voucher not found"
          },
          "409": {
            "description": "Voucher code already exists"
          },
          "429": {
            "description": "Too many requests",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RateLimitError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete a voucher",
        "tags": [
          "Vouchers"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Voucher ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Voucher deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Voucher not found"
          },
          "429": {
            "description": "Too many requests",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RateLimitError"
                }
              }
            }
          }
        }
      }
    }
  }
}